+++ kamehouse-groot/public/kame-house-groot/kamehouse-groot/js/kamehouse-shell.js
@@ -30,6 +30,7 @@ function KameHouseShell() {
       setScriptExecutingScriptOutput(scriptName, args, executeOnDockerHost);
       kameHouse.logger.info("Executing script : " + scriptName + " with args : '" + args + "' executeOnDockerHost: " + executeOnDockerHost);
       const config = kameHouse.http.getConfig();
+      config.timeout = 600;
       kameHouse.plugin.debugger.http.get(config, EXECUTE_API, kameHouse.http.getUrlEncodedHeaders(), params,
         (responseBody, responseCode, responseDescription, responseHeaders) => updateScriptOutput(responseBody, responseCode, responseDescription, responseHeaders, successCallback),
         (responseBody, responseCode, responseDescription, responseHeaders) => updateScriptOutputError(responseBody, responseCode, responseDescription, responseHeaders, errorCallback));